[refactor] fetch data server side
