+++ src/app/[lang]/utils/api-helpers.ts
@@ -1,3 +1,5 @@
+import { fetchAPI } from "./fetch-api";
+
 export function getStrapiURL(path = '', urlParamsObject = {}) {
     return `${process.env.NEXT_PUBLIC_STRAPI_API_URL || 'http://localhost:1337'}${path}`;
 }
@@ -15,3 +17,26 @@ export function formatDate(dateString: string) {
     const options: Intl.DateTimeFormatOptions = { year: 'numeric', month: 'long', day: 'numeric' };
     return date.toLocaleDateString('en-US', options);
 }
+
+export const fetchAboutPageData = async () => {
+    const path = `/about`;
+    const urlParamsObject = {
+        'populate[aboutHero][populate][title]': '*',
+        'populate[aboutHero][populate][heroImage]': '*',
+        'populate[twoColAbout][populate][title]': '*',
+        'populate[twoColAbout][populate][description]': '*',
+        'populate[twoColAbout][populate][content]': '*',
+        'populate[twoColAbout][populate][image]': '*',
+        'populate[twoColAbout][populate][flip]': '*',
+    };
+    return fetchAPI(path, urlParamsObject);
+};
+
+export const fetchContactPageData = async () => {
+    const path = `/contact`;
+    const urlParamsObject = {
+        'populate[contactSlider][populate][title]': '*',
+        'populate[contactSlider][populate][sliderImages]': '*',
+    };
+    return fetchAPI(path, urlParamsObject);
+};
